data: Move file review state into package
File review persistence records selected review context and exposes safety
checks for command flows through a single data module.

The file review domain is gaining records, fingerprint, freshness, and
action command helpers. Keeping the state store as a prefixed peer would
leave one domain split across two namespace shapes.

This commit begins the file review data package by moving the state store
into data.file_review.state and retargeting source imports to that package
path.

A following tests commit will retarget the base test imports before the
rest of the local series is replayed.
