+++ src/git_stage_batch/commands/discard.py
@@ -76,7 +76,7 @@ from ..data.hunk_tracking import (
     snapshot_selected_change_state,
     stream_live_git_diff,
 )
-from ..data.file_review_state import (
+from ..data.file_review.state import (
     FileReviewAction,
     clear_last_file_review_state_if_file_matches,
     finish_review_scoped_line_action,
