+++ src/git_stage_batch/commands/discard_from.py
@@ -30,7 +30,7 @@ from ..core.text_lifecycle import (
     normalized_text_change_type,
     selected_text_discard_change_type,
 )
-from ..data.file_review_state import (
+from ..data.file_review.state import (
     FileReviewAction,
     resolve_batch_source_action_scope,
 )
