+++ src/git_stage_batch/commands/include_from.py
@@ -46,7 +46,7 @@ from ..core.text_lifecycle import (
     normalized_text_change_type,
     selected_text_target_change_type,
 )
-from ..data.file_review_state import (
+from ..data.file_review.state import (
     FileReviewAction,
     finish_review_scoped_line_action,
     resolve_batch_source_action_scope,
