+++ src/git_stage_batch/data/file_review/__init__.py
@@ -1,0 +1,1 @@
+"""File review persistence and validation state."""
